backend/requirements.txt
Django
django-cors-headers
djangorestframework
gunicorn
Pillow
pytz
